refactor(Participant) : 딜러와 플레이어 사이에 발생하는 의존성을 제거
