+++ src/main/java/model/Participant.java
@@ -1,0 +1,37 @@
+package model;
+
+import java.util.ArrayList;
+import java.util.List;
+import model.dto.Card;
+import model.dto.PlayerResult;
+
+public class Participant {
+
+    private final String CARD_DUPLICATE_ERROR = "덱에 중복된 카드가 있습니다.";
+
+    private final PlayerName name;
+    private Integer score = 0;
+    private final List<Card> deck = new ArrayList<>();
+
+    public Participant(PlayerName name) {
+        this.name = name;
+    }
+
+    public PlayerResult getResult() {
+        return new PlayerResult(name, List.copyOf(deck), score);
+    }
+
+    public void addCard(Card card) {
+        validateCardDuplicate(card);
+        deck.add(card);
+        //점수 갱신 (메서드는 분리)
+    }
+
+    private void validateCardDuplicate(Card card) {
+        if(deck.contains(card)) {
+            throw new IllegalArgumentException(CARD_DUPLICATE_ERROR);
+        }
+    }
+
+
+}
