+++ src/main/java/model/PlayerName.java
@@ -11,17 +11,10 @@ public class PlayerName {
 
     private void validate(String name) {
         validateNameIsNotBlank(name);
-        validateNameIsNotAllowed(name);
     }
 
     private void validateNameIsNotBlank(String name) {
         if(name.isBlank()) {
-            throw new IllegalArgumentException("");
-        }
-    }
-
-    private void validateNameIsNotAllowed(String name) {
-        if(name.equals("딜러")) {
             throw new IllegalArgumentException("");
         }
     }
