+++ src/main/java/model/dto/PlayerResult.java
@@ -1,5 +1,6 @@
 package model.dto;
 
 import java.util.List;
+import model.PlayerName;
 
-public record PlayerResult(String name, List<Card> deck, Integer score) {}
+public record PlayerResult(PlayerName name, List<Card> deck, Integer score) {}
